requirements.txt
pylast
python-dotenv
lxml
beautifulsoup4>=4.4.0